+++ bin/ci/scripts/install_dependencies.sh
@@ -2,4 +2,4 @@
 
 composer self-update
 composer install
-composer require doctrine/couchdb:dev-master phpfastcache/phpssdb:~1.1.0 predis/predis:~1.1.0 mongodb/mongodb:^1.9
+composer require doctrine/couchdb:dev-master phpfastcache/phpssdb:~1.1.0 predis/predis:~1.1.0 mongodb/mongodb:^1.9 triagens/arangodb:^3.8